refactor(server): consolidate external providers
